Moved source files from src/lib to src.
Categorisation into lib and bin is actually unnecessary.
First, I know if a project I created is a lib or an executable project.
And second, even in an executable project with additional libs,
there can be no doubt which files are libs since I name the executable main.ts.

Furthermore, it simplifies creation of new projects
since there is less to change in package.json and tsconfig.json.
E. g.:
 - package.json for a lib project: "main" and "types" will always be
   "dist/main.js" and "dist/main.d.ts", respectively.
 - In an executable project with additional libs, a lib can be imported
   from main.js via path './myLib', instead of '../lib/myLib'.